refactor(msgmover): Cleanup legacy relays
Legacy relays are no longer supported
